+++ compass-geolocation-mobile/src/iosMain/kotlin/dev/jordond/compass/geolocation/mobile/internal/Mapper.kt
@@ -4,9 +4,9 @@ import dev.jordond.compass.Altitude
 import dev.jordond.compass.Azimuth
 import dev.jordond.compass.Coordinates
 import dev.jordond.compass.Location
+import dev.jordond.compass.Priority
 import dev.jordond.compass.Speed
 import dev.jordond.compass.geolocation.PermissionState
-import dev.jordond.compass.geolocation.Priority
 import kotlinx.cinterop.ExperimentalForeignApi
 import kotlinx.cinterop.useContents
 import platform.CoreLocation.CLAuthorizationStatus
