+++ compass-permissions-mobile/src/androidMain/AndroidManifest.xml
