+++ compass-permissions-mobile/src/androidMain/kotlin/dev/jordond/compass/permissions/mobile/internal/PermissionController.kt
@@ -1,13 +1,13 @@
-package dev.jordond.compass.geolocation.mobile.internal
+package dev.jordond.compass.permissions.mobile.internal
 
 import android.Manifest
 import android.content.Context
 import android.content.pm.PackageManager
 import androidx.core.content.ContextCompat
-import dev.jordond.compass.geolocation.PermissionState
-import dev.jordond.compass.geolocation.Priority
-import dev.jordond.compass.geolocation.mobile.internal.activity.ActivityProvider
+import dev.jordond.compass.Priority
+import dev.jordond.compass.permissions.PermissionState
 import dev.jordond.compass.permissions.exception.PermissionMissingException
+import dev.jordond.compass.permissions.mobile.internal.activity.ActivityProvider
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
