+++ compass-permissions/src/commonMain/kotlin/dev/jordond/compass/permissions/LocationPermissionController.kt
@@ -1,0 +1,4 @@
+package dev.jordond.compass.permissions
+
+public interface LocationPermissionController {
+}
